Refactor common mouse events utils
