+++ packages/jbrowse-plugin-apollo/src/util/mouseEventsUtils.ts
@@ -1,4 +1,5 @@
 import { type AnnotationFeature } from '@apollo-annotation/mst'
+import { type LinearGenomeViewModel } from '@jbrowse/plugin-linear-genome-view'
 
 type MinEdge = 'min'
 type MaxEdge = 'max'
@@ -110,4 +111,35 @@ export function getPropagatedLocationChanges(
     return shrinkFeatures(feature, newLocation, edge, shrinkParent)
   }
   return expandFeatures(feature, newLocation, edge)
+}
+
+/** extended information about the position of the mouse on the canvas, including the refName, bp, and displayedRegion number */
+export interface MousePosition {
+  x: number
+  y: number
+  refName: string
+  bp: number
+  regionNumber: number
+  feature?: AnnotationFeature
+}
+
+export type MousePositionWithFeature = Required<MousePosition>
+
+export function isMousePositionWithFeature(
+  mousePosition: MousePosition,
+): mousePosition is MousePositionWithFeature {
+  return 'feature' in mousePosition
+}
+
+export function getMousePosition(
+  event: React.MouseEvent,
+  lgv: LinearGenomeViewModel,
+): MousePosition {
+  const canvas = event.currentTarget
+  const { clientX, clientY } = event
+  const { left, top } = canvas.getBoundingClientRect()
+  const x = clientX - left
+  const y = clientY - top
+  const { coord: bp, index: regionNumber, refName } = lgv.pxToBp(x)
+  return { x, y, refName, bp, regionNumber }
 }
